+++ R/defaults.R
@@ -66,6 +66,8 @@ opts_chunk = new_defaults(list(
 
   warning = TRUE, error = TRUE, message = TRUE,
 
+  render = evaluate:::render,
+
   ref.label = NULL, child = NULL, engine = 'R', split = FALSE, include = TRUE, purl = TRUE
 
 ))